refactor: reorganize imports and enhance type definitions for clarity; add new graph and symbol types
